rename ContentCardTotalCostOfFeature file
