+++ components/vault/detailsSection/ContentCardTotalCostOfFeature.tsx
